move secured authors method to authorrestcontroller to make progam more neater
